Brought small cosmetic adjustments to script
The saving of amplitude matrices has been improved such that we can now really recover actual matrices.

Furthermore, an additional macro to align qubit pair is defined to ease readability in code.

On the post processing side, warning will be raised instead of errors when a line of plot is not available due to lack of data (e.g. if log XEB did yield only nan but not linear XEB).