Corrected multi qubit with CZ and added XEBResult.from_data method
Refactoring was done to enable the possibility to generate a XEBResult class from existing data that has been saved through the DataHandler previously.

Moreover, the simulation backend was adjusted such that the basis gates are consistent.

A new argument data_folder_name was added to the XEBConfig class such that one can choose which name to give to the saved data file.

Coupling map is now an attribute of the XEBConfig, as it can be inferred directly from qubits and qubit pairs.
XEBConfig also has a new method as_dict, converting it easily to data fields that are saved into a json file.

An extra optional field was added to the XEB.run method, enabling the user to specify a custom QuantumMachinesManager instance in case they want to simulate the program on the OPX Cloud simulator.