Automatically saving everything in same DataHandler at result gen
Saving data was changed in such a way that everything is now saved under the same folder within the XEBResult class. This save will occur only if should_save_data of the XEBConfig is set to True.
Limitation, it cannot save the singularities for now because of the format of this array that is not serializable.
Retrieving data should now also be doable from past data.